pull assert_doc_lists_equal and assert_docs_equal into a mixin
